Conditional (#128)
* Conditional mocking (#127)

* feat: conditional mocking and other enhancements

* Add conditional mocking functionality
* Add response functions returning string support
* Add enableMocks and disableMocks functions
* Add optional response function input and init arguments

* feat(types): add PR#114 changes

* doc(README): fix confusing documentation on conditional mocking

* test: correct timeout test

* feat: add do/dont mock and add only/never "If" suffix

* test: add complex test examples

* fix: incorporate changes from review to "once"

* fix: type linter errors

* Change implementation to use options object

* Revert changes for once

* Revert readme changes

* Remove error log

* Cleanup after refactor

* Add abort mocking and simplified typescript

* Updating readme with changes from API

* Cleanup and dependency update

Updated all deps to their most recent supported versions.  The TypeScript version of the index.d.ts file needed to be updated from 2.3 to 3.0 because Jest 23.x and higher requires TypeScript 3.0. Lowering the Jest version to 22.x (which supports typescript 2.3), causes a Type incompatibility because of a change in the jest.MockInstance generic argument definition.  Raising the Jest version to 24.x causes errors as it requires TypeScript 3.1 and higher.

Removed the .babelrc file and all babel related dependencies. Babel was only being used for a small amount of ES6 syntax used in the test cases and was complicating maintenance of the dependencies (and greatly increasing the amount of dependencies downloaded).  Removal only required 4 lines be changed in the tests.

Changed scripts to use yarn instead of npm run.

Fix error in abort script

* Removing .idea files

* Incorporate abort branch

* Removing IntelliJ files